refactor(core): move path security helper out of lib
