+++ polylogue/sources/parsers/base_models.py
@@ -8,8 +8,8 @@ from pydantic import BaseModel, Field, field_validator
 
 from polylogue.archive.conversation.branch_type import BranchType
 from polylogue.archive.message.types import MessageType
+from polylogue.core.security import sanitize_path as _sanitize_path_helper
 from polylogue.lib.roles import Role
-from polylogue.lib.security import sanitize_path as _sanitize_path_helper
 from polylogue.types import ContentBlockType, Provider
